refactor: use prompts in code
